+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EAS project config [[#63](https://github.com/pLTdwaxn/Caroushow/issues/63)]
 - Prebuild iOS with Expo New Architecture [#82](https://github.com/pLTdwaxn/Caroushow/issues/82)
 - Ability to cycle through Aspect Ratios [#87](https://github.com/pLTdwaxn/Caroushow/issues/87)
+- Complex Redux action logic [#89](https://github.com/pLTdwaxn/Caroushow/issues/89)
 
 ### Changed